pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for monolithic app with FastAPI and Supabase"
authors = ["matias cox <mcox5@uc.cl>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
supabase = "^2.10.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
boto3 = "^1.35.68"
python-multipart = "^0.0.17"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
